Crate fc_rpc

Modules

format

frontier_backend_client

pending

Structs

Debug

Debug API implementation.

Eth

Eth API implementation.

EthBlockDataCacheTask

Manage LRU caches for block data and their transaction statuses. These are large and take a lot of time to fetch from the database. Storing them in an LRU cache will allow to reduce database accesses when many subsequent requests are related to the same blocks.

EthDevSigner

EthFilter

EthPubSub

Eth pub-sub API implementation.

EthTask

EthereumSubIdProvider

Net

Net API implementation.

RuntimeApiStorageOverride

A storage override for runtimes that use runtime API.

SchemaV1StorageOverride

A storage override for runtimes that use schema v1.

SchemaV1StorageOverrideRef

A storage override reference for runtimes that use schema v1.

SchemaV2StorageOverride

A storage override for runtimes that use schema v2.

SchemaV2StorageOverrideRef

A storage override reference for runtimes that use schema v2.

SchemaV3StorageOverride

A storage override for runtimes that use schema v3.

SchemaV3StorageOverrideRef

A storage override for runtimes that use schema v3.

StorageOverrideHandler

A storage override for runtimes that use different ethereum schema.

StorageQuerier

A useful utility for querying storage.

TxPool

Web3

Web3 API implementation.

Enums

EthereumTransaction

Traits

DebugApiServer

Server trait implementation for the DebugApi RPC API.

EstimateGasAdapter

Allow to adapt a request for estimate_gas. Can be used to estimate gas of some contracts using a different function in the case the normal gas estimation doesn’t work.

EthApiServer

Server trait implementation for the EthApi RPC API.

EthConfig

EthFilterApiServer

Server trait implementation for the EthFilterApi RPC API.

EthPubSubApiServer

Server trait implementation for the EthPubSubApi RPC API.

EthSigner

A generic Ethereum signer.

NetApiServer

Server trait implementation for the NetApi RPC API.

StorageOverride

This trait is used to obtain Ethereum-related data.

TxPoolApiServer

Server trait implementation for the TxPoolApi RPC API.

Web3ApiServer

Server trait implementation for the Web3Api RPC API.

Functions

err

internal_err

internal_err_with_data

public_key