fc_rpc

Module frontier_backend_client

Structs

SystemAccountId20StorageOverride

Implements a default runtime storage override. It assumes that the balances and nonces are stored in pallet system.account, and have nonce: Index = u32 for and free: Balance = u128. Uses IdentityAddressMapping for the address.

SystemAccountId32StorageOverride

Implements a runtime storage override. It assumes that the balances and nonces are stored in pallet system.account, and have nonce: Index = u32 for and free: Balance = u128. USes HashedAddressMapping for the address.

Functions

is_canon

load_hash

load_transactions

native_block_id