Trait fc_rpc::DebugApiServer

pub trait DebugApiServer: Sized + Send + Sync + 'static {
    // Required methods
    fn raw_header<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash
    ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn raw_block<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash
    ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: H256
    ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn raw_receipts<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Bytes>, Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn bad_blocks(&self, number: BlockNumberOrHash) -> Result<Vec<()>, Error>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the DebugApi RPC API.

Required Methods

fn raw_header<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>where 'life0: 'async_trait, Self: 'async_trait,

Returns an RLP-encoded header with the given number or hash.

fn raw_block<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>where 'life0: 'async_trait, Self: 'async_trait,

Returns an RLP-encoded block with the given number or hash.

fn raw_transaction<'life0, 'async_trait>( &'life0 self, hash: H256 ) -> Pin<Box<dyn Future<Output = Result<Option<Bytes>, Error>> + Send + 'async_trait>>where 'life0: 'async_trait, Self: 'async_trait,

Returns a EIP-2718 binary-encoded transaction with the given hash.

fn raw_receipts<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash ) -> Pin<Box<dyn Future<Output = Result<Vec<Bytes>, Error>> + Send + 'async_trait>>where 'life0: 'async_trait, Self: 'async_trait,

Returns an array of EIP-2718 binary-encoded receipts with the given number of hash.

fn bad_blocks(&self, number: BlockNumberOrHash) -> Result<Vec<()>, Error>

Returns an array of recent bad blocks that the client has seen on the network.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.