Trait fc_rpc::EthApi

pub trait EthApi: 'static + Send + Sync {

    fn protocol_version(&self) -> Result<u64, Error>;
    fn syncing(&self) -> Result<SyncStatus, Error>;
    fn author(&self) -> Result<H160, Error>;
    fn accounts(&self) -> Result<Vec<H160, Global>, Error>;
    fn block_number(&self) -> Result<U256, Error>;
    fn chain_id(&self) -> Result<Option<U64>, Error>;
    fn block_by_hash(
        &self, 
        hash: H256, 
        full: bool
    ) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>;
    fn block_by_number(
        &self, 
        number: BlockNumber, 
        full: bool
    ) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>;
    fn block_transaction_count_by_hash(
        &self, 
        hash: H256
    ) -> Result<Option<U256>, Error>;
    fn block_transaction_count_by_number(
        &self, 
        number: BlockNumber
    ) -> Result<Option<U256>, Error>;
    fn block_uncles_count_by_hash(&self, hash: H256) -> Result<U256, Error>;
    fn block_uncles_count_by_number(
        &self, 
        number: BlockNumber
    ) -> Result<U256, Error>;
    fn uncle_by_block_hash_and_index(
        &self, 
        hash: H256, 
        index: Index
    ) -> Result<Option<Rich<Block>>, Error>;
    fn uncle_by_block_number_and_index(
        &self, 
        number: BlockNumber, 
        index: Index
    ) -> Result<Option<Rich<Block>>, Error>;
    fn transaction_by_hash(
        &self, 
        hash: H256
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;
    fn transaction_by_block_hash_and_index(
        &self, 
        hash: H256, 
        index: Index
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;
    fn transaction_by_block_number_and_index(
        &self, 
        number: BlockNumber, 
        index: Index
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;
    fn transaction_receipt(
        &self, 
        hash: H256
    ) -> Pin<Box<dyn Future<Output = Result<Option<Receipt>, Error>> + Send + 'static, Global>>;
    fn balance(
        &self, 
        address: H160, 
        number: Option<BlockNumber>
    ) -> Result<U256, Error>;
    fn storage_at(
        &self, 
        address: H160, 
        index: U256, 
        number: Option<BlockNumber>
    ) -> Result<H256, Error>;
    fn transaction_count(
        &self, 
        address: H160, 
        number: Option<BlockNumber>
    ) -> Result<U256, Error>;
    fn code_at(
        &self, 
        address: H160, 
        number: Option<BlockNumber>
    ) -> Result<Bytes, Error>;
    fn call(
        &self, 
        request: CallRequest, 
        number: Option<BlockNumber>
    ) -> Result<Bytes, Error>;
    fn estimate_gas(
        &self, 
        request: CallRequest, 
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'static, Global>>;
    fn gas_price(&self) -> Result<U256, Error>;
    fn fee_history(
        &self, 
        block_count: U256, 
        newest_block: BlockNumber, 
        reward_percentiles: Option<Vec<f64, Global>>
    ) -> Result<FeeHistory, Error>;
    fn max_priority_fee_per_gas(&self) -> Result<U256, Error>;
    fn is_mining(&self) -> Result<bool, Error>;
    fn hashrate(&self) -> Result<U256, Error>;
    fn work(&self) -> Result<Work, Error>;
    fn submit_hashrate(&self, hashrate: U256, id: H256) -> Result<bool, Error>;
    fn submit_work(
        &self, 
        nonce: H64, 
        pow_hash: H256, 
        mix_digest: H256
    ) -> Result<bool, Error>;
    fn send_transaction(
        &self, 
        request: TransactionRequest
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>;
    fn send_raw_transaction(
        &self, 
        bytes: Bytes
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>;

    fn to_delegate<M>(self) -> IoDelegate<Self, M>
    where
        M: Metadata,
    { ... }

}

Eth rpc interface.

Required Methods

fn protocol_version(&self) -> Result<u64, Error>

Returns protocol version encoded as a string (quotes are necessary).

fn syncing(&self) -> Result<SyncStatus, Error>

Returns an object with data about the sync status or false. (wtf?)

fn author(&self) -> Result<H160, Error>

Returns block author.

fn accounts(&self) -> Result<Vec<H160, Global>, Error>

Returns accounts list.

fn block_number(&self) -> Result<U256, Error>

Returns highest block number.

fn chain_id(&self) -> Result<Option<U64>, Error>

Returns the chain ID used for transaction signing at the current best block. None is returned if not available.

fn block_by_hash(
&self,
hash: H256,
full: bool
) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>

Returns block with given hash.

fn block_by_number(
&self,
number: BlockNumber,
full: bool
) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>

Returns block with given number.

fn block_transaction_count_by_hash(
&self,
hash: H256
) -> Result<Option<U256>, Error>

Returns the number of transactions in a block with given hash.

fn block_transaction_count_by_number(
&self,
number: BlockNumber
) -> Result<Option<U256>, Error>

Returns the number of transactions in a block with given block number.

fn block_uncles_count_by_hash(&self, hash: H256) -> Result<U256, Error>

Returns the number of uncles in a block with given hash.

fn block_uncles_count_by_number(
&self,
number: BlockNumber
) -> Result<U256, Error>

Returns the number of uncles in a block with given block number.

fn uncle_by_block_hash_and_index(
&self,
hash: H256,
index: Index
) -> Result<Option<Rich<Block>>, Error>

Returns an uncles at given block and index.

fn uncle_by_block_number_and_index(
&self,
number: BlockNumber,
index: Index
) -> Result<Option<Rich<Block>>, Error>

Returns an uncles at given block and index.

fn transaction_by_hash(
&self,
hash: H256
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Get transaction by its hash.

fn transaction_by_block_hash_and_index(
&self,
hash: H256,
index: Index
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Returns transaction at given block hash and index.

fn transaction_by_block_number_and_index(
&self,
number: BlockNumber,
index: Index
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Returns transaction by given block number and index.

fn transaction_receipt(
&self,
hash: H256
) -> Pin<Box<dyn Future<Output = Result<Option<Receipt>, Error>> + Send + 'static, Global>>

Returns transaction receipt by transaction hash.

fn balance(
&self,
address: H160,
number: Option<BlockNumber>
) -> Result<U256, Error>

Returns balance of the given account.

fn storage_at(
&self,
address: H160,
index: U256,
number: Option<BlockNumber>
) -> Result<H256, Error>

Returns content of the storage at given address.

fn transaction_count(
&self,
address: H160,
number: Option<BlockNumber>
) -> Result<U256, Error>

Returns the number of transactions sent from given address at given time (block number).

fn code_at(
&self,
address: H160,
number: Option<BlockNumber>
) -> Result<Bytes, Error>

Returns the code at given address at given time (block number).

fn call(
&self,
request: CallRequest,
number: Option<BlockNumber>
) -> Result<Bytes, Error>

Call contract, returning the output data.

fn estimate_gas(
&self,
request: CallRequest,
number: Option<BlockNumber>
) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'static, Global>>

Estimate gas needed for execution of given contract.

fn gas_price(&self) -> Result<U256, Error>

Returns current gas_price.

fn fee_history(
&self,
block_count: U256,
newest_block: BlockNumber,
reward_percentiles: Option<Vec<f64, Global>>
) -> Result<FeeHistory, Error>

Introduced in EIP-1159 for getting information on the appropriate priority fee to use.

fn max_priority_fee_per_gas(&self) -> Result<U256, Error>

Introduced in EIP-1159, a Geth-specific and simplified priority fee oracle. Leverages the already existing fee history cache.

fn is_mining(&self) -> Result<bool, Error>

Returns true if client is actively mining new blocks.

fn hashrate(&self) -> Result<U256, Error>

Returns the number of hashes per second that the node is mining with.

fn work(&self) -> Result<Work, Error>

Returns the hash of the current block, the seedHash, and the boundary condition to be met.

fn submit_hashrate(&self, hashrate: U256, id: H256) -> Result<bool, Error>

Used for submitting mining hashrate.

fn submit_work(
&self,
nonce: H64,
pow_hash: H256,
mix_digest: H256
) -> Result<bool, Error>

Used for submitting a proof-of-work solution.

fn send_transaction(
&self,
request: TransactionRequest
) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>

Sends transaction; will block waiting for signer to return the transaction hash.

fn send_raw_transaction(
&self,
bytes: Bytes
) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>

Sends signed transaction, returning its hash.

Provided Methods

fn to_delegate<M>(self) -> IoDelegate<Self, M> where
M: Metadata,

Create an IoDelegate, wiring rpc calls to the trait methods.