fc_rpc

Trait NetApiServer

pub trait NetApiServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn version(&self) -> Result<String, ErrorObject<'static>>;
    fn peer_count(&self) -> Result<PeerCount, ErrorObject<'static>>;
    fn is_listening(&self) -> Result<bool, ErrorObject<'static>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the NetApi RPC API.

Required Methods

fn version(&self) -> Result<String, ErrorObject<'static>>

Returns protocol version.

fn peer_count(&self) -> Result<PeerCount, ErrorObject<'static>>

Returns number of peers connected to node.

fn is_listening(&self) -> Result<bool, ErrorObject<'static>>

Returns true if client is actively listening for network connections. Otherwise false.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.