fc_rpc

Trait Web3ApiServer

pub trait Web3ApiServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn client_version(&self) -> Result<String, ErrorObject<'static>>;
    fn sha3(&self, input: Bytes) -> Result<H256, ErrorObject<'static>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the Web3Api RPC API.

Required Methods

fn client_version(&self) -> Result<String, ErrorObject<'static>>

Returns current client version.

fn sha3(&self, input: Bytes) -> Result<H256, ErrorObject<'static>>

Returns sha3 of the given data

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.