Trait fc_rpc::EthApiServer

pub trait EthApiServer: 'static + Send + Sync {

    fn protocol_version(&self) -> Result<u64, Error>;

    fn syncing(&self) -> Result<SyncStatus, Error>;

    fn hashrate(&self) -> Result<U256, Error>;

    fn author(&self) -> Result<H160, Error>;

    fn is_mining(&self) -> Result<bool, Error>;

    fn chain_id(&self) -> Result<Option<U64>, Error>;

    fn gas_price(&self) -> Result<U256, Error>;

    fn accounts(&self) -> Result<Vec<H160, Global>, Error>;

    fn block_number(&self) -> Result<U256, Error>;

    fn balance(&self, H160, Option<BlockNumber>) -> Result<U256, Error>;

    fn storage_at(&self, H160, U256, Option<BlockNumber>) -> Result<H256, Error>;

    fn block_by_hash(
        &self, 
        H256, 
        bool
    ) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>;

    fn block_by_number(
        &self, 
        BlockNumber, 
        bool
    ) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>;

    fn transaction_count(
        &self, 
        H160, 
        Option<BlockNumber>
    ) -> Result<U256, Error>;

    fn block_transaction_count_by_hash(
        &self, 
        H256
    ) -> Result<Option<U256>, Error>;

    fn block_transaction_count_by_number(
        &self, 
        BlockNumber
    ) -> Result<Option<U256>, Error>;

    fn block_uncles_count_by_hash(&self, H256) -> Result<U256, Error>;

    fn block_uncles_count_by_number(&self, BlockNumber) -> Result<U256, Error>;

    fn code_at(&self, H160, Option<BlockNumber>) -> Result<Bytes, Error>;

    fn send_transaction(
        &self, 
        TransactionRequest
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>;

    fn send_raw_transaction(
        &self, 
        Bytes
    ) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>;

    fn call(&self, CallRequest, Option<BlockNumber>) -> Result<Bytes, Error>;

    fn estimate_gas(
        &self, 
        CallRequest, 
        Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'static, Global>>;

    fn transaction_by_hash(
        &self, 
        H256
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;

    fn transaction_by_block_hash_and_index(
        &self, 
        H256, 
        Index
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;

    fn transaction_by_block_number_and_index(
        &self, 
        BlockNumber, 
        Index
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>;

    fn transaction_receipt(
        &self, 
        H256
    ) -> Pin<Box<dyn Future<Output = Result<Option<Receipt>, Error>> + Send + 'static, Global>>;

    fn uncle_by_block_hash_and_index(
        &self, 
        H256, 
        Index
    ) -> Result<Option<Rich<Block>>, Error>;

    fn uncle_by_block_number_and_index(
        &self, 
        BlockNumber, 
        Index
    ) -> Result<Option<Rich<Block>>, Error>;

    fn logs(
        &self, 
        Filter
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Log, Global>, Error>> + Send + 'static, Global>>;

    fn work(&self) -> Result<Work, Error>;

    fn submit_work(&self, H64, H256, H256) -> Result<bool, Error>;

    fn submit_hashrate(&self, U256, H256) -> Result<bool, Error>;

    fn fee_history(
        &self, 
        block_count: U256, 
        newest_block: BlockNumber, 
        reward_percentiles: Option<Vec<f64, Global>>
    ) -> Result<FeeHistory, Error>;

    fn to_delegate<M>(self) -> IoDelegate<Self, M>
    where
        M: Metadata,
    { ... }

}

Eth rpc interface.

Required methods

fn protocol_version(&self) -> Result<u64, Error>

Returns protocol version encoded as a string (quotes are necessary).

fn syncing(&self) -> Result<SyncStatus, Error>

Returns an object with data about the sync status or false. (wtf?)

fn hashrate(&self) -> Result<U256, Error>

Returns the number of hashes per second that the node is mining with.

fn author(&self) -> Result<H160, Error>

Returns block author.

fn is_mining(&self) -> Result<bool, Error>

Returns true if client is actively mining new blocks.

fn chain_id(&self) -> Result<Option<U64>, Error>

Returns the chain ID used for transaction signing at the current best block. None is returned if not available.

fn gas_price(&self) -> Result<U256, Error>

Returns current gas_price.

fn accounts(&self) -> Result<Vec<H160, Global>, Error>

Returns accounts list.

fn block_number(&self) -> Result<U256, Error>

Returns highest block number.

fn balance(&self, H160, Option<BlockNumber>) -> Result<U256, Error>

Returns balance of the given account.

fn storage_at(&self, H160, U256, Option<BlockNumber>) -> Result<H256, Error>

Returns content of the storage at given address.

fn block_by_hash(
&self,
H256,
bool
) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>

Returns block with given hash.

fn block_by_number(
&self,
BlockNumber,
bool
) -> Pin<Box<dyn Future<Output = Result<Option<Rich<Block>>, Error>> + Send + 'static, Global>>

Returns block with given number.

fn transaction_count(&self, H160, Option<BlockNumber>) -> Result<U256, Error>

Returns the number of transactions sent from given address at given time (block number).

fn block_transaction_count_by_hash(&self, H256) -> Result<Option<U256>, Error>

Returns the number of transactions in a block with given hash.

fn block_transaction_count_by_number(
&self,
BlockNumber
) -> Result<Option<U256>, Error>

Returns the number of transactions in a block with given block number.

fn block_uncles_count_by_hash(&self, H256) -> Result<U256, Error>

Returns the number of uncles in a block with given hash.

fn block_uncles_count_by_number(&self, BlockNumber) -> Result<U256, Error>

Returns the number of uncles in a block with given block number.

fn code_at(&self, H160, Option<BlockNumber>) -> Result<Bytes, Error>

Returns the code at given address at given time (block number).

fn send_transaction(
&self,
TransactionRequest
) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>

Sends transaction; will block waiting for signer to return the transaction hash.

fn send_raw_transaction(
&self,
Bytes
) -> Pin<Box<dyn Future<Output = Result<H256, Error>> + Send + 'static, Global>>

Sends signed transaction, returning its hash.

fn call(&self, CallRequest, Option<BlockNumber>) -> Result<Bytes, Error>

Call contract, returning the output data.

fn estimate_gas(
&self,
CallRequest,
Option<BlockNumber>
) -> Pin<Box<dyn Future<Output = Result<U256, Error>> + Send + 'static, Global>>

Estimate gas needed for execution of given contract.

fn transaction_by_hash(
&self,
H256
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Get transaction by its hash.

fn transaction_by_block_hash_and_index(
&self,
H256,
Index
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Returns transaction at given block hash and index.

fn transaction_by_block_number_and_index(
&self,
BlockNumber,
Index
) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>, Error>> + Send + 'static, Global>>

Returns transaction by given block number and index.

fn transaction_receipt(
&self,
H256
) -> Pin<Box<dyn Future<Output = Result<Option<Receipt>, Error>> + Send + 'static, Global>>

Returns transaction receipt by transaction hash.

fn uncle_by_block_hash_and_index(
&self,
H256,
Index
) -> Result<Option<Rich<Block>>, Error>

Returns an uncles at given block and index.

fn uncle_by_block_number_and_index(
&self,
BlockNumber,
Index
) -> Result<Option<Rich<Block>>, Error>

Returns an uncles at given block and index.

fn logs(
&self,
Filter
) -> Pin<Box<dyn Future<Output = Result<Vec<Log, Global>, Error>> + Send + 'static, Global>>

Returns logs matching given filter object.

fn work(&self) -> Result<Work, Error>

Returns the hash of the current block, the seedHash, and the boundary condition to be met.

fn submit_work(&self, H64, H256, H256) -> Result<bool, Error>

Used for submitting a proof-of-work solution.

fn submit_hashrate(&self, U256, H256) -> Result<bool, Error>

Used for submitting mining hashrate.

fn fee_history(
&self,
block_count: U256,
newest_block: BlockNumber,
reward_percentiles: Option<Vec<f64, Global>>
) -> Result<FeeHistory, Error>

Introduced in EIP-1159 for getting information on the appropiate priority fee to use.

Provided methods

fn to_delegate<M>(self) -> IoDelegate<Self, M> where
M: Metadata,

Create an IoDelegate, wiring rpc calls to the trait methods.