Trait fc_rpc::StorageOverride

pub trait StorageOverride<Block: BlockT> {
    fn account_code_at(
        &self, 
        block: &BlockId<Block>, 
        address: H160
    ) -> Option<Vec<u8>>;

    fn storage_at(
        &self, 
        block: &BlockId<Block>, 
        address: H160, 
        index: U256
    ) -> Option<H256>;

    fn current_block(&self, block: &BlockId<Block>) -> Option<EthereumBlock>;

    fn current_receipts(&self, block: &BlockId<Block>) -> Option<Vec<Receipt>>;

    fn current_transaction_statuses(
        &self, 
        block: &BlockId<Block>
    ) -> Option<Vec<TransactionStatus>>;
}

Something that can fetch Ethereum-related data. This trait is quite similar to the runtime API, and indeed oe implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-ethereum’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Required methods

fn account_code_at(
&self,
block: &BlockId<Block>,
address: H160
) -> Option<Vec<u8>>

For a given account address, returns pallet_evm::AccountCodes.

fn storage_at(
&self,
block: &BlockId<Block>,
address: H160,
index: U256
) -> Option<H256>

For a given account address and index, returns pallet_evm::AccountStorages.

fn current_block(&self, block: &BlockId<Block>) -> Option<EthereumBlock>

Return the current block.

fn current_receipts(&self, block: &BlockId<Block>) -> Option<Vec<Receipt>>

Return the current receipt.

fn current_transaction_statuses(
&self,
block: &BlockId<Block>
) -> Option<Vec<TransactionStatus>>

Return the current transaction status.